refactor: redefine structure
