+++ src/json-rpc/create-session-memory-store.1.ts
@@ -1,0 +1,9 @@
+import { SessionMemoryStore } from "./session-memory-store.js";
+
+/**
+ * Factory function to create a session memory store.
+ * @param sessionId - The session ID for the store
+ * @returns A new SessionMemoryStore instance
+ */
+export const sessionMemoryStore = (sessionId: string) =>
+  new SessionMemoryStore(sessionId);
