+++ src/json-rpc/index.ts
@@ -1,9 +1,7 @@
+export type { JsonRpcErrorResponse } from "./types/json-rpc-error-response.js";
+export type { JsonRpcHandler } from "./types/json-rpc-handler.js";
+export type { JsonRpcRequest } from "./types/json-rpc-request.js";
+export type { JsonRpcResponse } from "./types/json-rpc-response.js";
+export type { JsonRpcResultResponse } from "./types/json-rpc-result-response.js";
+export { JsonRpcError } from "./json-rpc-error.js";
 export { JsonRpcDispatcher } from "./json-rpc-dispatcher.js";
-export type {
-  JsonRpcRequest,
-  JsonRpcResponse,
-  JsonRpcResultResponse,
-  JsonRpcErrorResponse,
-  JsonRpcHandler,
-} from "./types.js";
-export { JsonRpcError } from "./types.js";
