+++ src/json-rpc/types/extract-validation-type.ts
@@ -1,0 +1,4 @@
+import type { Validation } from "./validation";
+
+export type ExtractValidationType<A> =
+  A extends Validation<infer U> ? U : unknown;
