+++ src/json-rpc/types/json-rpc-error-response.ts
@@ -1,0 +1,9 @@
+export interface JsonRpcErrorResponse<T = any> {
+  jsonrpc: "2.0";
+  id: string | number;
+  error: {
+    code: number;
+    message: string;
+    data?: T;
+  };
+}
