+++ src/json-rpc/types/json-rpc-event.ts
@@ -1,0 +1,3 @@
+export type JsonRpcEvent = {
+  httpRequest?: Request;
+};
