+++ src/json-rpc/types/json-rpc-request.ts
@@ -1,0 +1,6 @@
+export interface JsonRpcRequest<T = any> {
+  jsonrpc: "2.0";
+  id: string | number;
+  method: string;
+  params?: T;
+}
