+++ src/json-rpc/types/json-rpc-result-response.ts
@@ -1,0 +1,5 @@
+export interface JsonRpcResultResponse<T = any> {
+  jsonrpc: "2.0";
+  id: string | number;
+  result: T;
+}
