+++ src/json-rpc/types/validation.ts
@@ -1,0 +1,3 @@
+import type { ZodValidation } from "./zod-validation";
+
+export type Validation<T> = ZodValidation<T>;
