+++ src/json-rpc/types/zod-validation.ts
@@ -1,0 +1,3 @@
+export type ZodValidation<T> = {
+  safeParse: (data: any) => { success: boolean; data?: T; error?: any };
+};
